Moved the test resources back to the test folder
