+++ src/test/java/no/ntnu/idata2304/group1/server/network/TCPListenerTest.java
@@ -12,14 +12,6 @@ public class TCPListenerTest {
         TCPListener listener = new TCPListener(6008,
                 TCPListener.class.getResource("TestKeys").getPath().toString().replace("%20", " "),
                 "123");
-        listener.run();
-    }
-
-    public static void main(String[] args) throws Exception {
-        TCPListener listener = new TCPListener(6008,
-                TCPListener.class.getResource("TestKeys").getPath().toString().replace("%20", " "),
-                "123");
-        listener.run();
     }
 }
