+++ src/test/resources/no/ntnu/idata2304/group1/server/README.md
@@ -3,3 +3,7 @@
 ## TestKeys for network
 
 The test keys contains a paired certificate for testing. This can be opened up by using the keytool CLI commands or other programs like KeyStore explorer. The password is set to be 123 (This is test keys and are not meant to be sent in production so there is no danger by noting the password here).
+
+## TestDB
+
+The TestDB contains test information like rooms, sensor and dummy logs.
